Program uses multiple threads to find values
Idea of the program is that it is provided with supplier interface whitch gives it hidden packages. Those contain values but not all of them, so we need to use a certain number of threads to search all those packages and add all the found values. When we did it we provide the supplier with this sum and are provided the next set of packages.
Note: Interfaces visible here were provided with the exercise I am not their author, but the program wont work without them.