Shop abstractions
A shop has its magazine, whitch contains products, and clients willing to buy them. When a product thet set client wants to buy isn't present in the magazine or is in not sufficient quantity, client will wait in a line. When the product becomes present all clients who can buy it will do it.